requirements.txt
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
click==8.2.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distlib==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.55.8
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
stack-data==0.6.3
tabula-py==2.10.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.1
virtualenv==20.32.0
wcwidth==0.2.13
Werkzeug==3.1.3
xlrd==2.0.2
